[hzh] feat: add no strike or spare

+++ src/main/java/BowlingGame.java
@@ -2,7 +2,7 @@ import java.util.Arrays;
 
 public class BowlingGame {
     public int countSource(int[] numberOfDown) {
-        int source = (int) Arrays.stream(numberOfDown).sum();
+        int source = Arrays.stream(numberOfDown).sum();
 
         return source;
     }
